fix: change code after PR rejected
