+++ packages/frontend/web/src/components/CocktailCard.tsx
@@ -4,7 +4,7 @@ export default function CocktailCard() {
   return (
     <div className='relative -top-8'>
       <div className='border-dark bg-card-pink absolute -top-3 left-10 z-30 m-auto my-20 h-[21rem] w-[18rem] rounded-sm border-[3px]' />
-      <div className='border-dark bg-pastel-brown absolute -top-6 left-6  m-auto my-20 h-[21rem] w-[18rem] rounded-sm border-[3px]' />
+      <div className='border-dark bg-pastel-brown absolute -top-6 left-6 m-auto my-20 h-[21rem] w-[18rem] rounded-sm border-[3px]' />
       <div className='border-dark bg-card-pink-dark absolute left-14 z-[50] my-20 h-[21rem] w-[18rem] rounded-sm border-[3px] uppercase'>
         <img
           src='cocktail-placeholder.png'
@@ -14,7 +14,6 @@ export default function CocktailCard() {
         <div>
           <p className='font-stroke text-light text-md mx-4 mt-3 text-center'>{`Cocktail's name`}</p>
           <div className='mt-2 flex justify-center'>
-            {' '}
             <StarRating starCount={5} />
           </div>
         </div>
