+++ packages/types/src/cocktail-detail.ts
@@ -1,0 +1,11 @@
+export interface CommentProps {
+  numberComment: number;
+}
+
+export interface StarRatingProps {
+  starCount: number;
+}
+
+export interface StarProps {
+  index: number;
+}
