+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './database';
+export * from './cocktail-detail';
 
 export type User = {
   name: string;
